fix: limit positions to those you have access to

+++ app/Http/Controllers/Admin/PositionController.php
@@ -4,7 +4,10 @@ namespace App\Http\Controllers\Admin;
 
 use App\Helpers\VatsimRating;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\PositionRequest;
+use App\Models\Area;
 use App\Models\Position;
+use App\Services\PositionService;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -12,65 +15,45 @@ use Illuminate\Validation\Rule;
 
 class PositionController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware(function ($request, $next) {
-            if (Auth::user()->isModeratorOrAbove()) {
-                return $next($request);
-            }
 
-            abort(403);
-        });
-    }
 
-    public function index(): View
+    public function index(PositionService $positionService): View
     {
-        $positions = Position::with('area')->get();
+        $this->authorize('viewAny', Position::class);
+
+        $positions = $positionService->getPositions();
         $ratings = collect(VatsimRating::cases())->filter(function ($rating) {
-            return !in_array($rating, [VatsimRating::INA, VatsimRating::SUS, VatsimRating::OBS, VatsimRating::SUP, VatsimRating::ADM]);
+            return !in_array($rating, VatsimRating::NOT_POSITION_RATINGS);
         });
+        $canCreate = Auth::user()->groups()->where('id', '<=', 2)->exists();
 
-        return view('admin.positions.index', compact('positions', 'ratings'));
+        return view('admin.positions.index', compact('positions', 'ratings', 'canCreate'));
     }
 
-    public function store(Request $request)
+    public function store(PositionRequest $request)
     {
-        $data = $request->validate([
-            'callsign' => 'required|unique:positions,callsign',
-            'name' => 'required',
-            'frequency' => 'required',
-            'fir' => 'required',
-            'rating' => ['required', 'integer', Rule::in(collect(VatsimRating::cases())->filter(function ($rating) {
-                return !in_array($rating, [VatsimRating::INA, VatsimRating::SUS, VatsimRating::OBS, VatsimRating::SUP, VatsimRating::ADM]);
-            })->pluck('value')->toArray())],
-            'area_id' => 'required|integer',
-        ]);
+        $this->authorize('create', new Position($request->all()));
 
-        $position = Position::create($data);
+        $position = Position::create($request->validated());
 
         return redirect()->route('positions.index')->with('success', 'Position ' . $position->callsign . ' created successfully.');
     }
 
-    public function update(Request $request, Position $position)
+    public function update(PositionRequest $request, Position $position)
     {
-        $data = $request->validate([
-            'callsign' => 'required|unique:positions,callsign,' . $position->id,
-            'name' => 'required',
-            'frequency' => 'required',
-            'fir' => 'required',
-            'rating' => ['required', 'integer', Rule::in(collect(VatsimRating::cases())->filter(function ($rating) {
-                return !in_array($rating, [VatsimRating::INA, VatsimRating::SUS, VatsimRating::OBS, VatsimRating::SUP, VatsimRating::ADM]);
-            })->pluck('value')->toArray())],
-            'area_id' => 'required|integer',
-        ]);
+        $tempPosition = new Position($request->all());
+        $tempPosition->area = Area::findOrFail($request->area_id);
+        $this->authorize('update', $tempPosition);
 
-        $position->update($data);
+        $position->update($request->validated());
 
         return redirect()->route('positions.index')->with('success', 'Position ' . $position->callsign . ' updated successfully.');
     }
 
     public function destroy(Position $position)
     {
+        $this->authorize('delete', $position);
+
         $position->delete();
 
         return redirect()->route('positions.index')->with('success', 'Position ' . $position->callsign . ' deleted successfully.');
